despensai/requirements.txt
django
djangorestframework
openai
python-dotenv
mysqlclient
pillow